fix: add bin config and fix run project
